feat:Trainer 페이지 분리작업
